add basic inventory creation

+++ src/NosCore.Database/Entities/ItemInstance.cs
@@ -20,7 +20,7 @@ namespace NosCore.Database.Entities
 
         #region Properties
 
-        public int Amount { get; set; }
+        public short Amount { get; set; }
 
         public virtual ICollection<BazaarItem> BazaarItem { get; set; }
